Separated UI logic in inbox
